Added a new possibility for implementating the clock selector
The current clock selector has a glitch when Baud CLK is low when the clock switches from TGL to BCLK. For a instant, CLKOUT will go high and then go low again. This might be long enough to trigger the shift register holding our data like we want, but it also might not be. I don't like it and I worry that it might cause issues, so I intend on removing it.